+++ frontend/src/components/index.ts
@@ -33,7 +33,7 @@ export * from 'components/DarkModeButton';
 export * from 'components/CommentItem';
 export * from 'components/FeedItem';
 export * from 'components/@shared/BottomSheet';
-export * from 'components/Dropdown';
+export * from 'components/@shared/Dropdown';
 export * from 'components/Bell';
 export * from 'components/SubscriptionButton';
 export * from 'components/CustomCycleTimeBottomSheet';
